server/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
passlib==1.7.2
protobuf==3.13.0
psycopg2==2.8.5
PyJWT==1.4.2
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
uuid==1.30
Werkzeug==1.0.1
